[ANDROSDK-1312-2] Tracker importer: manage deleted items
